Fixed memory leaks in ssw_cpp in case of non-default translation matrix.
The leak was reported by valgrind as:
128 bytes in 1 blocks are definitely lost in loss record 1 of 1
   at 0x4C294E0: operator new[](unsigned long) (in vgpreload_memcheck-amd64-linux.so)
   by 0x404CD8: StripedSmithWaterman::Aligner::Aligner(signed char const*, int const&, signed char const*, int const&) (ssw_cpp.cpp:288)

Also fixed a leak in Rebuild().

Got rid of matrix_built_ -- check translation_matrix_ pointer for NULL
instead.
Got rid of default_matrix_ too as the actual logic using it was wrong --
instead always perform a full copy of the translation matrix (even in
case of default one).